Move StatefulReaderProps interface
